+++ src/wrapper/descriptors/field_desc.rs
@@ -6,24 +6,24 @@ use crate::{
     JNIEnv,
 };
 
-impl<'a, 'c, T, U, V> Desc<'a, JFieldID<'a>> for (T, U, V)
+impl<'a, 'c, T, U, V> Desc<'a, JFieldID> for (T, U, V)
 where
     T: Desc<'a, JClass<'c>>,
     U: Into<JNIString>,
     V: Into<JNIString>,
 {
-    fn lookup(self, env: &JNIEnv<'a>) -> Result<JFieldID<'a>> {
+    fn lookup(self, env: &JNIEnv<'a>) -> Result<JFieldID> {
         env.get_field_id(self.0, self.1, self.2)
     }
 }
 
-impl<'a, 'c, T, U, V> Desc<'a, JStaticFieldID<'a>> for (T, U, V)
+impl<'a, 'c, T, U, V> Desc<'a, JStaticFieldID> for (T, U, V)
 where
     T: Desc<'a, JClass<'c>>,
     U: Into<JNIString>,
     V: Into<JNIString>,
 {
-    fn lookup(self, env: &JNIEnv<'a>) -> Result<JStaticFieldID<'a>> {
+    fn lookup(self, env: &JNIEnv<'a>) -> Result<JStaticFieldID> {
         env.get_static_field_id(self.0, self.1, self.2)
     }
 }