+++ src/wrapper/objects/jstaticmethodid.rs
@@ -1,19 +1,23 @@
 use crate::sys::jmethodID;
 
-/// Wrapper around `sys::jmethodid` that implements `Send` + `Sync` since method IDs
-/// are valid across threads (not tied to a `JNIEnv`). There is no lifetime associated
-/// with these since they aren't garbage collected like objects and their lifetime
-/// is not implicitly connected with the scope in which they are queried.
+/// Wrapper around [`jmethodID`] that implements `Send` + `Sync` since method IDs
+/// are valid across threads (not tied to a `JNIEnv`).
+///
+/// There is no lifetime associated with these since they aren't garbage
+/// collected like objects and their lifetime is not implicitly connected with
+/// the scope in which they are queried.
 ///
 /// It matches C's representation of the raw pointer, so it can be used in any
-/// of the extern function argument positions that would take a `jmethodid`.
+/// of the extern function argument positions that would take a [`jmethodID`].
 ///
 /// # Safety
 ///
-/// According to the JNI spec method IDs may be released when the corresponding class is
-/// unloaded. Since this constraint can't be encoded as a Rust lifetime,
-/// and to avoid the excessive cost of having every Method ID be associated with
-/// a global reference to the corresponding class then it is the developers
+/// According to the JNI spec method IDs may be invalidated when the
+/// corresponding class is unloaded.
+///
+/// Since this constraint can't be encoded as a Rust lifetime, and to avoid the
+/// excessive cost of having every Method ID be associated with a global
+/// reference to the corresponding class then it is the developers
 /// responsibility to ensure they hold some class reference for the lifetime of
 /// cached method IDs.
 #[repr(transparent)]